+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Date sorting is now using descending permalink as tie breaker instead of ascending to match date sorting
 - Cache result of `get_url(..., cachebust=true)`
 - Do not error when setting heading ids in content with a Tera error
-- Properly fix taxonomy slugs merging
+- Properly merge taxonomies terms when needed after slugification
 
 ## 0.23.3 (2026-08-11)