+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.23.4 (unreleased)
+## 0.23.4 (2026-08-20)
 
 - Fix another `get_url` bug adding a trailing slash for non default language
 - Fix `index.<default_language>.md` paths collision